+++ middleman-core/features/step_definitions/page_layout_steps.rb
@@ -4,12 +4,3 @@ Given /^page "([^\"]*)" has layout "([^\"]*)"$/ do |url, layout|
     page(url, :layout => layout.to_sym)
   }
 end
-
-Given /^"([^\"]*)" with_layout block has layout "([^\"]*)"$/ do |url, layout|
-  @initialize_commands ||= []
-  @initialize_commands << lambda {
-    with_layout(layout.to_sym) do
-      page(url)
-    end
-  }
-end