+++ middleman-core/fixtures/auto-js-app/config.rb
@@ -1,11 +1,9 @@
-with_layout false do
-  %w{
-    /auto-js.html
-    /auto-js
-    /auto-js/
-    /auto-js/auto-js.html
-    /auto-js/sub/auto-js.html
-  }.each do |path|
-    page path
-  end
+%w{
+  /auto-js.html
+  /auto-js
+  /auto-js/
+  /auto-js/auto-js.html
+  /auto-js/sub/auto-js.html
+}.each do |path|
+  page path, layout: false
 end