+++ middleman-core/fixtures/basic-data-app/config.rb
@@ -1,4 +1,2 @@
-with_layout false do
-  page "/data.html"
-  page "/data3.html"
-end
+page "/data.html", layout: false
+page "/data3.html", layout: false