+++ middleman-core/fixtures/capture-html-app/config.rb
@@ -1,7 +1,5 @@
 require "slim"
 
-with_layout :capture_html do
-  page "/capture_html_erb.html"
-  page "/capture_html_haml.html"
-  page "/capture_html_slim.html"
-end
+page "/capture_html_erb.html", layout: :capture_html
+page "/capture_html_haml.html", layout: :capture_html
+page "/capture_html_slim.html", layout: :capture_html