+++ middleman-core/fixtures/content-for-app/config.rb
@@ -1,7 +1,5 @@
 require "slim"
 
-with_layout :content_for do
-  page "/content_for_erb.html"
-  page "/content_for_haml.html"
-  page "/content_for_slim.html"
-end
+page "/content_for_erb.html", layout: :content_for
+page "/content_for_haml.html", layout: :content_for
+page "/content_for_slim.html", layout: :content_for