+++ middleman-core/fixtures/generator-test/config.rb
@@ -22,11 +22,6 @@
 #
 # With alternative layout
 # page "/path/to/file.html", :layout => :otherlayout
-#
-# A path which all have the same layout
-# with_layout :admin do
-#   page "/admin/*"
-# end
 
 # Proxy (fake) files
 # page "/this-page-has-no-template.html", :proxy => "/template-file.html" do