+++ middleman-core/fixtures/large-build-app/config.rb
@@ -1,3 +1,1 @@
-with_layout false do
-  page "/spaces in file.html"
-end
+page "/spaces in file.html", layout: false