+++ middleman-core/fixtures/page-classes-app/config.rb
@@ -1,4 +1,2 @@
-with_layout false do
-  page "/sub1/page-classes.html"
-  page "/sub1/sub2/page-classes.html"
-end
+page "/sub1/page-classes.html", layout: false
+page "/sub1/sub2/page-classes.html", layout: false